+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Data `gest` and `struct`.
 
+* `separate` and `sep_by` arguments in `predict_gam()` allow the user to separate variables in the model that were created with `interaction()`.
+
 ## Fixed
 
 * Error when predicting bivariate smooths (`s/te/ti()`, `fs/re` basis functions) where only the first variable was returned internally.